New version of app without phone numbers.

+++ src/io/forsta/ccsm/database/ContactDb.java
@@ -31,10 +31,12 @@ public class ContactDb extends DbBase {
   public static final String EMAIL = "email";
   public static final String NUMBER = "number";
   public static final String USERNAME = "username";
+  public static final String AVATAR = "avatar";
   public static final String UID = "uid";
   public static final String TAGID = "tagid";
   public static final String SLUG = "slug";
   public static final String ORGID = "orgid";
+  public static final String ORGSLUG = "org_slug";
   public static final String DATE = "date";
   public static final String TSREGISTERED = "tsregistered";
 
@@ -45,13 +47,15 @@ public class ContactDb extends DbBase {
       EMAIL + ", " +
       NUMBER + ", " +
       USERNAME + ", " +
+      AVATAR + ", " +
       UID + ", " +
       TAGID + ", " +
       SLUG + ", " +
       ORGID + ", " +
+      ORGSLUG + ", " +
       DATE + ", " +
       TSREGISTERED + " integer default 0, " +
-      "CONSTRAINT item_number_unique UNIQUE (" + NUMBER + ")" +
+      "CONSTRAINT item_number_unique UNIQUE (" + UID + ")" +
       ")";
 
   public static String[] allColumns = {
@@ -60,10 +64,12 @@ public class ContactDb extends DbBase {
       EMAIL,
       NUMBER,
       USERNAME,
+      AVATAR,
       UID,
       TAGID,
       SLUG,
       ORGID,
+      ORGSLUG,
       DATE,
       TSREGISTERED
   };
@@ -78,11 +84,25 @@ public class ContactDb extends DbBase {
 
   public Cursor getContactByAddress(String address) {
     try {
-      return getRecords(TABLE_NAME, null, NUMBER + " = ?", new String[] {address}, NUMBER);
+      return getRecords(TABLE_NAME, null, UID + " = ?", new String[] {address}, UID);
     } catch (Exception e) {
       e.printStackTrace();
     }
     return null;
+  }
+
+  public ForstaUser getUserByAddress(String address) {
+    ForstaUser user = null;
+    try {
+      Cursor cursor = getRecords(TABLE_NAME, null, UID + " = ?", new String[] {address}, UID);
+      if (cursor != null && cursor.moveToNext()) {
+        user = new ForstaUser(cursor);
+      }
+      cursor.close();
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+    return user;
   }
 
   public HashMap<String, String> getContactSlugs() {
@@ -90,7 +110,7 @@ public class ContactDb extends DbBase {
     try {
       Cursor c = getRecords(TABLE_NAME, allColumns, TSREGISTERED + "=1", null, USERNAME);
       while (c.moveToNext()) {
-        contacts.put(c.getString(c.getColumnIndex(USERNAME)), c.getString(c.getColumnIndex(NUMBER)));
+        contacts.put(c.getString(c.getColumnIndex(UID)), c.getString(c.getColumnIndex(SLUG)) + ":" + c.getString(c.getColumnIndex(ORGSLUG)));
       }
       c.close();
     } catch (Exception e) {
@@ -99,13 +119,19 @@ public class ContactDb extends DbBase {
     return contacts;
   }
 
-  public HashMap<String, ForstaRecipient> getContactRecipients() {
+  public HashMap<String, ForstaRecipient> getContactRecipients(String localOrgSlug) {
     HashMap<String, ForstaRecipient> contacts = new HashMap<>();
     try {
-      Cursor c = getRecords(TABLE_NAME, allColumns, TSREGISTERED + "=1", null, USERNAME);
+      Cursor c = getRecords(TABLE_NAME, null, TSREGISTERED + "=1", null, USERNAME);
       while (c.moveToNext()) {
-        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGID)));
-        contacts.put(c.getString(c.getColumnIndex(USERNAME)), recipient);
+        String slug = c.getString(c.getColumnIndex(SLUG));
+        String orgSlug = c.getString(c.getColumnIndex(ContactDb.ORGSLUG));
+        if (!TextUtils.equals(orgSlug, localOrgSlug)) {
+          slug += ":" + orgSlug;
+        }
+        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGSLUG)));
+
+        contacts.put(slug, recipient);
       }
       c.close();
     } catch (Exception e) {
@@ -126,6 +152,20 @@ public class ContactDb extends DbBase {
       e.printStackTrace();
     }
     return numbers;
+  }
+
+  public Set<String> getAddresses() {
+    Set<String> addresses = new HashSet<>();
+    try {
+      Cursor c = getRecords(TABLE_NAME, allColumns, null, null, UID);
+      while (c.moveToNext()) {
+        addresses.add(c.getString(c.getColumnIndex(UID)));
+      }
+      c.close();
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+    return addresses;
   }
 
   public Map<String, String> getUids() {
@@ -162,7 +202,7 @@ public class ContactDb extends DbBase {
     try {
       Cursor c = getRecords(TABLE_NAME, allColumns, TSREGISTERED + "=1", null, NAME);
       while (c.moveToNext()) {
-        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGID)));
+        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGSLUG)));
         recipients.add(recipient);
       }
       c.close();
@@ -172,7 +212,7 @@ public class ContactDb extends DbBase {
     return recipients;
   }
 
-  public void updateUsers(List<ForstaUser> users) {
+  public void updateUsers(List<ForstaUser> users, boolean removeExisting) {
     SQLiteDatabase db = mDbHelper.getWritableDatabase();
     Set<String> forstaUids = new HashSet<>();
     Map<String, String> uids = getUids();
@@ -186,8 +226,10 @@ public class ContactDb extends DbBase {
         values.put(ContactDb.TAGID, user.tag_id);
         values.put(ContactDb.SLUG, user.slug);
         values.put(ContactDb.UID, user.uid);
+        values.put(ContactDb.AVATAR, user.avatar);
         values.put(ContactDb.NAME, user.name);
         values.put(ContactDb.ORGID, user.org_id);
+        values.put(ContactDb.ORGSLUG, user.org_slug);
         values.put(ContactDb.NUMBER, user.phone);
         values.put(ContactDb.USERNAME, user.username);
         values.put(ContactDb.EMAIL, user.email);
@@ -205,17 +247,18 @@ public class ContactDb extends DbBase {
     finally {
       db.endTransaction();
     }
-    db.beginTransaction();
-    try {
-      // Now remove entries that are no longer valid.
-      for (String uid : uids.keySet()) {
-        db.delete(TABLE_NAME, UID + "=?", new String[] { uid });
+    if (removeExisting) {
+      db.beginTransaction();
+      try {
+        // Now remove entries that are no longer valid.
+        for (String uid : uids.keySet()) {
+          db.delete(TABLE_NAME, UID + "=?", new String[] { uid });
+        }
+        db.setTransactionSuccessful();
+      } finally {
+        db.endTransaction();
       }
-      db.setTransactionSuccessful();
-    } finally {
-      db.endTransaction();
     }
-
     db.close();
   }
 
@@ -228,14 +271,32 @@ public class ContactDb extends DbBase {
     db.update(TABLE_NAME, values, ID + "=?", new String[] { user.id });
   }
 
-  public void setActiveForstaNumbers(List<ContactTokenDetails> activeTokens) {
+  public void setActiveForstaAddresses(List<ContactTokenDetails> activeTokens, Set<String> eligibleAddresses) {
     SQLiteDatabase db = mDbHelper.getWritableDatabase();
     // This could be done with a update TABLE_NAME set TSREGISTERED = 1 where number in (1,2,3)
     for (ContactTokenDetails token : activeTokens) {
-      String number = token.getNumber();
+      String address = token.getNumber();
       ContentValues values = new ContentValues();
       values.put(TSREGISTERED, true);
-      db.update(TABLE_NAME, values, NUMBER + "=?", new String[] { number });
+      db.update(TABLE_NAME, values, UID + "=?", new String[] { address });
+    }
+
+    for (String address : eligibleAddresses) {
+      ContentValues values = new ContentValues();
+      values.put(TSREGISTERED, false);
+      db.update(TABLE_NAME, values, UID + "=?", new String[] { address });
+    }
+    db.close();
+  }
+
+  public void setActiveForstaAddresses(List<ContactTokenDetails> activeTokens) {
+    SQLiteDatabase db = mDbHelper.getWritableDatabase();
+    // This could be done with a update TABLE_NAME set TSREGISTERED = 1 where number in (1,2,3)
+    for (ContactTokenDetails token : activeTokens) {
+      String address = token.getNumber();
+      ContentValues values = new ContentValues();
+      values.put(TSREGISTERED, true);
+      db.update(TABLE_NAME, values, UID + "=?", new String[] { address });
     }
     db.close();
   }
@@ -244,7 +305,7 @@ public class ContactDb extends DbBase {
     String queryFilter = TSREGISTERED + " = 1";
     String[] queryValues = null;
     if (filter != null && filter.length() > 0) {
-      queryFilter += " AND (" + NAME + " LIKE ? OR " + NUMBER + " LIKE ?)";
+      queryFilter += " AND (" + NAME + " LIKE ? OR " + SLUG + " LIKE ?)";
       queryValues = new String[] { "%" + filter + "%", "%" + filter + "%" };
     }
 
@@ -285,11 +346,11 @@ public class ContactDb extends DbBase {
 
     String query = "";
     String queryNumbers = TextUtils.join("','", numbers);
-    query = NUMBER + " IN ('" + queryNumbers + "')";
+    query = UID + " IN ('" + queryNumbers + "')";
     try {
       Cursor c = getRecords(TABLE_NAME, allColumns, query, null, ORGID + ", " + NAME);
       while (c.moveToNext()) {
-        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGID)));
+        ForstaRecipient recipient = new ForstaRecipient(c.getString(c.getColumnIndex(ContactDb.NAME)), c.getString(c.getColumnIndex(ContactDb.NUMBER)), c.getString(c.getColumnIndex(ContactDb.USERNAME)), c.getString(c.getColumnIndex(ContactDb.UID)), c.getString(c.getColumnIndex(ContactDb.ORGSLUG)));
         recipients.add(recipient);
       }
       c.close();
